MB-65665 - xdcrDiffer basic docker support
For the near term, these changes will allow the xdcrDiffer
itself to be compiled into a docker image and run as a docker
container. The readme has been updated to include a basic
instruction set for performing this.

Note that this functionality will not be a long term support.
See the MB itself for more info

Change-Id: I5e024948d2e044a590a1618de7e72a5f65f9605e
Reviewed-on: https://review.couchbase.org/c/xdcrDiffer/+/224867
Tested-by: Neil Huang <neil.huang@couchbase.com>
Reviewed-by: Sumukh Bhat <sumukh.bhat@couchbase.com>